Add JavaScript Code for BHB: The BlockHeads Book
I am adding the JavaScript Code File for BHB: The BlockHeads Book to my GitHub repository. This is the code that I have worked on for the app so far, and I will continue to work on it (and my other projects) in the future! If you have any issues, requests, or any advice, then that is 100% appreciated! Thanks! :)